Made files for classes, added colorize gem
