+++ Gemfile
@@ -1,0 +1,3 @@
+source 'https://rubygems.org'
+
+gem 'colorize', '~> 0.8.1'
